Linux BTEQ export to csv
BTEQ job reads databasenames and objects  names from a parameter file generates and export script for each object and uses sed to remove the whitespace for csv.  This job has been used to produce sample datasets from a large numnber of tables from a number of databases. BTEQ does not do csv out of box but outputs a header record which TPT and other Teradata utilities do not.